changed the whole darkmode logic and added it inside app.jsx so that we can pass it through different components instead of having it stuck in a far toggleswitch button
